update repository name and refactor
