+++ .travis.yml
@@ -29,7 +29,7 @@ jobs:
             - codecov
 after_success:
   - docker login -e $DOCKER_EMAIL -u $DOCKER_USERNAME -p $DOCKER_PASSWORD
-  - export REPOSITORY=ralston3/kademlia
+  - export REPOSITORY=ralston3/liaa
   - docker build -f ./Dockerfile -t $REPOSITORY:$COMMIT .
   - docker tag $REPOSITORY:$COMMIT $REPOSITORY:$TRAVIS_BRANCH
   - docker tag $REPOSITORY:$COMMIT $REPOSITORY:travis-$TRAVIS_BUILD_NUMBER
