+++ docs/conf.py
@@ -49,8 +49,8 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'Kademlia'
-copyright = u'2018, Brian Muller'
+project = u'Liaa'
+copyright = u'2019, Rashad Alston'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -58,8 +58,8 @@ copyright = u'2018, Brian Muller'
 #
 # The short X.Y version.
 sys.path.insert(0, os.path.abspath('..'))
-import kademlia
-version = kademlia.__version__
+import liaa
+version = liaa.__version__
 # The full version, including alpha/beta/rc tags.
 release = version
