+++ examples/simple_peer.py
@@ -15,8 +15,8 @@ import asyncio
 import sys
 import getopt
 
-from kademlia.network import Server
-from kademlia.utils import ArgsParser
+from liaa.network import Server
+from liaa.utils import ArgsParser
 
 
 def usage():
@@ -31,7 +31,7 @@ def main():
 	handler = logging.StreamHandler()
 	formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
 	handler.setFormatter(formatter)
-	log = logging.getLogger('kademlia')
+	log = logging.getLogger('liaa')
 	log.addHandler(handler)
 	log.setLevel(logging.DEBUG)
