+++ pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-addopts = -vv --cov-report term-missing --cov kademlia
+addopts = -vv --cov-report term-missing --cov liaa
